+++ annotations/nocedit.exe/pseudocode/src/engine/2d.c/computeOutcode_FUN_00402c10.c
@@ -2,11 +2,11 @@
 // Address: 00402c10
 // Address Range: [[00402c10, 00402c41]]
 // Convention: __cdecl
-// Signature: byte engine_2d.c_computeOutcode_FUN_00402c10(int x, int y, int x_min, int y_min, int x_max, int y_max)
+// Signature: int engine_2d.c_computeOutcode_FUN_00402c10(int x, int y, int x_min, int y_min, int x_max, int y_max)
 
 #include "nocturne.h"
 
-byte __cdecl
+int __cdecl
 engine_2d_c_computeOutcode_FUN_00402c10(int x,int y,int x_min,int y_min,int x_max,int y_max)
 
 {
@@ -22,5 +22,5 @@ engine_2d_c_computeOutcode_FUN_00402c10(int x,int y,int x_min,int y_min,int x_ma
   if (x < x_min) {
     bVar1 = bVar1 | 8;
   }
-  return bVar1;
+  return (uint)bVar1;
 }
