+++ annotations/nocedit.exe/pseudocode/src/shape/design.c/showFacetPartEditor_FUN_0045f1d0.c
@@ -114,8 +114,7 @@ void __cdecl shape_design_c_showFacetPartEditor_FUN_0045f1d0(void)
       local_68.y = 0.0;
       local_68.z = 0.0;
       engine_3d_c_processCameraRelativePoint_FUN_004037e0(&local_68);
-      engine_matrix_c_matrixPushAndTransform_FUN_0050cee0
-                ((ushort)local_58,(ushort)local_3c,(ushort)local_38,0,0,0);
+      engine_matrix_c_matrixPushAndTransform_FUN_0050cee0(local_58,local_3c,local_38,0,0,0);
       if (g_WireframeMode != 0) {
         dVar9 = crt_math_c_round_FUN_005fe6b0((double)local_20);
         iVar4 = (int)ROUND(dVar9);
